Update README.md
Big Bang Theory is a sitcom about 7 friends discussing about their daily lives, relationships and a little bit of science; Interstellar discuss about time and space where an astronaut travels in space and get traps in space and time. Avengers: End Game is a superhero movie discussing about saving the rest of population in the whole universe that's been erased by the villain.